+++ app/libraries/Setups/Views/Sidebar.php
@@ -10,6 +10,11 @@ final class Sidebar extends AbstractSetup
     public function run()
     {
         \add_action('jentil_before_before_footer', [$this, 'load']);
+        \add_action('jentil_after_after_header', [$this, 'openContentWrapTag']);
+        \add_action(
+            'jentil_before_before_footer',
+            [$this, 'closeContentWrapTag']
+        );
     }
 
     /**
@@ -22,5 +27,33 @@ final class Sidebar extends AbstractSetup
         }
 
         $this->app->utilities->loader->loadPartial('sidebar');
+    }
+
+    /**
+     * @action jentil_after_after_header
+     */
+    public function openContentWrapTag()
+    {
+        if ($this->app->utilities->postTypeTemplate->isPageBuilder() ||
+            'columns-1' === $this->app->utilities->page->layout->column()
+        ) {
+            return;
+        }
+
+        echo '<div id="content-wrap">';
+    }
+
+    /**
+     * @action jentil_before_before_footer
+     */
+    public function closeContentWrapTag()
+    {
+        if ($this->app->utilities->postTypeTemplate->isPageBuilder() ||
+            'columns-1' === $this->app->utilities->page->layout->column()
+        ) {
+            return;
+        }
+
+        echo '</div><!-- #content-wrap -->';
     }
 }