+++ app/partials/footer.php
@@ -1,8 +1,7 @@
 <?php
 declare (strict_types = 1);
 
-        \do_action('jentil_after_after_content'); ?>
-    </div><!-- #content -->
-</div><!-- #content-wrap -->
+    \do_action('jentil_after_after_content'); ?>
+</div><!-- #content -->
 
 <?php \Jentil()->utilities->loader->loadPartial('footer', 'mini');