+++ app/partials/sidebar.php
@@ -11,11 +11,9 @@ if ('columns-1' === ($column = \Jentil()->utilities->page->layout->column()) &&
  * Primary Sidebar
  */
 if (\is_active_sidebar($id = \Jentil()->setups['Sidebars\Primary']->id)) { ?>
-    <div id="primary-sidebar-wrap" class="sidebar-wrap">
-        <aside id="primary-sidebar" class="site-sidebar widget-area" itemscope itemtype="http://schema.org/WPSideBar">
-            <?php \dynamic_sidebar($id); ?>
-        </aside><!-- #primary -->
-    </div>
+    <aside id="primary-sidebar" class="site-sidebar widget-area" itemscope itemtype="http://schema.org/WPSideBar">
+        <?php \dynamic_sidebar($id); ?>
+    </aside><!-- #primary -->
 <?php }
 
 if ('columns-3' !== $column &&
@@ -28,9 +26,7 @@ if ('columns-3' !== $column &&
  * Secondary sidebar
  */
 if (\is_active_sidebar($id = \Jentil()->setups['Sidebars\Secondary']->id)) { ?>
-    <div id="secondary-sidebar-wrap" class="sidebar-wrap">
-        <aside id="secondary-sidebar" class="site-sidebar widget-area" itemscope itemtype="http://schema.org/WPSideBar">
-            <?php \dynamic_sidebar($id); ?>
-        </aside><!-- #secondary -->
-    </div>
+    <aside id="secondary-sidebar" class="site-sidebar widget-area" itemscope itemtype="http://schema.org/WPSideBar">
+        <?php \dynamic_sidebar($id); ?>
+    </aside><!-- #secondary -->
 <?php }